Yeniden Ergenekon

+++ src/Ergenekon.Application/World/Queries/GetCountries/GetCountriesQuery.cs
@@ -4,9 +4,9 @@ using MediatR;
 
 namespace Ergenekon.Application.World.Queries.GetCountries;
 
-public record GetCountriesQuery : IRequest<List<LookupDto1<byte>>>;
+public record GetCountriesQuery : IRequest<List<LookupDto1<ushort>>>;
 
-public class GetCountriesQueryHandler : IRequestHandler<GetCountriesQuery, List<LookupDto1<byte>>>
+public class GetCountriesQueryHandler : IRequestHandler<GetCountriesQuery, List<LookupDto1<ushort>>>
 {
     private readonly IWorldService _worldService;
 
@@ -15,7 +15,7 @@ public class GetCountriesQueryHandler : IRequestHandler<GetCountriesQuery, List<
         _worldService = worldService;
     }
 
-    public async Task<List<LookupDto1<byte>>> Handle(GetCountriesQuery request, CancellationToken cancellationToken)
+    public async Task<List<LookupDto1<ushort>>> Handle(GetCountriesQuery request, CancellationToken cancellationToken)
     {
         return await _worldService.GetAllCountriesAsync(cancellationToken);
     }
