+++ src/Ergenekon.Application/World/Queries/GetProvincesByCountryId/GetProvincesByCountryIdQuery.cs
@@ -4,7 +4,7 @@ using MediatR;
 
 namespace Ergenekon.Application.World.Queries.GetProvincesByCountryId;
 
-public record GetProvincesByCountryIdQuery(byte CountryId) : IRequest<List<LookupDto1<ushort>>>;
+public record GetProvincesByCountryIdQuery(ushort CountryId) : IRequest<List<LookupDto1<ushort>>>;
 
 public class GetProvincesByCountryIdQueryHandler : IRequestHandler<GetProvincesByCountryIdQuery, List<LookupDto1<ushort>>>
 {
